chore: Updated input fields
- Combine them into a single file
- Removed use of ref (since we never really used them)
  - I think this was added for supporting the inputs exported by Gorhom Bottom Sheets, but we discarded their use since the `editable` prop didn't work
