+++ mobile/src/components/new/Form/index.ts
@@ -1,10 +1,8 @@
 export { Button } from "./Button";
 export { IconButton } from "./IconButton";
+export { NumericInput, TextInput } from "./Input";
 export { Legend, LegendItem } from "./Legend";
-export { NumericInput } from "./NumericInput";
 export { ProgressBar } from "./ProgressBar";
 export { Ripple } from "./Ripple";
 export { Checkbox, Radio } from "./Selection";
 export { Switch } from "./Switch";
-export { TextInput } from "./TextInput";
-export { useInputRef } from "./useInputRef";
